Adding modalfor QR Code in account
