+++ package.json
@@ -13,7 +13,9 @@
     "eject": "react-native-scripts eject",
     "android": "react-native-scripts android",
     "ios": "react-native-scripts ios",
-    "test": "jest"
+    "test": "jest",
+    "debug": "open \"rndebugger://set-debugger-loc?host=localhost&port=19001\""
+
   },
   "jest": {
     "preset": "jest-expo"
